+++ lib/ecto_mysql_extras/queries/total_index_size.ex
@@ -21,7 +21,7 @@ defmodule EctoMySQLExtras.TotalIndexSize do
     """
     /* ECTO_MYSQL_EXTRAS: #{info().title} */
 
-    SELECT CAST(COALESCE(SUM(ROUND(INDEX_LENGTH / 1024 / 1024)),0) AS UNSIGNED) AS `size`
+    SELECT CAST(COALESCE(SUM(INDEX_LENGTH),0) AS UNSIGNED) AS `size`
     FROM information_schema.tables
     WHERE TABLE_SCHEMA = DATABASE();
     """